refact MainRoutes extract SearchRoutes reduce fn lines
